+++ DSCResources/MSFT_ScheduledTask/MSFT_ScheduledTask.psm1
@@ -1449,7 +1449,10 @@ function Test-TargetResource
 
     Write-Verbose -Message ($script:localizedData.TestingDscParameterStateMessage)
 
-    return Test-DscParameterState -CurrentValues $currentValues -DesiredValues $desiredValues
+    return Test-DscParameterState `
+        -CurrentValues $currentValues `
+        -DesiredValues $desiredValues `
+        -Verbose:$VerbosePreference
 }
 
 <#